datatype

+++ pcmr/featurizers/hf.py
@@ -43,7 +43,7 @@ class HuggingFaceFeaturizerMixin(BatchSizeMixin):
     def __call__(self, smis: Iterable[str]) -> np.ndarray:
         return torch.stack(
             [H[0, self.CLASS_TOKEN_IDX, :] for H in self.fe(smis, batch_size=self.batch_size)]
-        ).numpy()
+        ).numpy().astype(float)
 
     def finetune(self, *splits: Iterable[tuple[Iterable[str], ArrayLike]]) -> Self:
         raise NotImplementedError
